Added Dealing Step
Added Dealing step+lots of debugging. Was unable to properly debug using VSCode, so went to printing...

Currently player hands will sort on initial draw of the 13 tiles. Should sort each time a tile is added as well. Attempted to use more iterative approaches to handling elements in a list, but ran into java.util.ConcurrentModificationException because I was iterating on a mutable list and trying to add an element to it at the same time.

I ended up using a traditional for i loop with a return after the element was added. I could use a couple other methods like saving the index, using a hash map, or using a linked list to save some time and space complexity.

Next will be starting the round and having the dealer draw the first tile. Then I'll add in discard logic, as well as checks for yaku and tenpai.